active Decoratorを導入 Viewに関するmethodを記載する
